Convert styling props definitions
